pyproject.toml
[project]
name = "xhs-mcp-silent"
version = "0.3.0"
description = "Silent Xiaohongshu CLI that reuses Chrome cookies on macOS without Playwright or Chromium automation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "curl-cffi>=0.10.0",
    "pycryptodome>=3.20.0",
    "PyExecJS>=1.5.1",
]

[project.scripts]
xhs-silent = "xhs_mcp_silent.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = [
    "xhs_mcp_silent/vendor/xhsvm.js",
]

[tool.hatch.build.targets.wheel]
packages = ["xhs_mcp_silent"]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]
